Refactored each component into the proper corresponding folder
Also removed unused react components
