+++ src/components/apparatusLists/ApparatusSelectedListBox.tsx
